requirements.txt
Flask==2.3.2
ibm-cloud-sdk-core==3.16.7
ibm-watson==7.0.0
python-dotenv==1.0.0
requests==2.31.0